Fix the Zephyr HxM Bluetooth checksum algorithm
When using RunnerUp with a Zephyr, I was only seeing very sporadic HR values:
5 in the space of 15 minutes, that sort of thing; and no periodicity.

I guessed the checksum code was broken and sure enough it was.

This patch is to fix the Zephyr HxM Bluetooth checksum. Specifically:
 - push the checksum code down to the ZephyrHRM class
 - rather than simply XOR-ing bytes together, use the algorithm specified in
   http://www.zephyranywhere.com/zephyr-labs/development-tools/hxm-bt/communications-packet-structure-overview/

TESTED=ran in Eclipse/ADT; confirmed no parseBuffer errors; saw HR values
on the connection console in the app.