Code changes to support requesting permissions dynamically
GPS is requested at App startup. (This is tricky to do when pressing Start button, the request is done asynchronously.)
A few corrections to handle GPS not started
If no GPS, a toast to inform the user that GPS is not on
Longer term, TabHost should be replaced with something AppCompat Theme setup, so a Material Design SnackBar can be used. Seem to require SDK 14, so not with Froyo/Gingerbread
The GPS can be set to work like other sensors, to be able to deactivate it

Also for read/write backups. There it would be natural to always request permission, but this do not work, so a toast there too.