Use AppCompat Theme
AppCompat theme is required for some newer elements in Support Design library like SnackBars
Using AppCompat also aligns the design with other apps,removes some of the dated look
Some overrides could be removed in code (I did not change the override in styles.xml)

The update slightly changes fonts, dividers and looks, the following "larger" changes are seen (may be more):
 * It is hard to have the Start/Feed/History/Settings with both icon and text. I choose the icons
 * Backgrounds were previously mostly black, but not all (for instance Advanced Settings were graded). Overrides removed, all gray now

I would like to use SnackBars to handle permissions
The branch is on top of the SDK update right now, as it is needed with SDK update. However, the commit could be applied to master too
Before doing that and testing on other than Marshmallow, I would like to have feedback if this can be accepted.

Longer term, the app design should be updated.
Maybe Light theme?
Use more Material Design, skipping the two levels of tabs and use a titlebar?