Update tab style to match previous setup
Both image and text is visible. KK and above have no real issues what I see. CheckBoxes have slightly different appearance than stock though.
Manage Workout button text is also visible in JB and earlier, but it is lost on updating. This seems to be related to the use of background for the buttons, no solution yet.