Prepare for Play Services update
No actual change. Propose it is merged so change if is available if testing for some reason - annoying to find these issues

Update Play Services requires applicationId to be set. Using a dynamic approach
The play version is not updated right now. No new api is needed what I know and updating just increases the size with 3MB (which is an issue on older devices).