Code changes to support requesting MM permissions dynamically
GPS is requested at App startup. If not granted a SnackBar will nag the user (unless not ask again is ticked). (This is tricky to do when pressing Start button, the request is done asynchronously.)
A few corrections to handle GPS not started
Longer term, the GPS can be set to work like other sensors, to be able to deactivate it

Also for read/write backups. There it would be natural to always request permission, but this do not work, so popup asking the user to manually activate.

Only popup for storage permissions
The request is not working there for some reason. (This could be improved with at least supplying a link to the settings).
This should be good enough for now.

Note that sharing activities and workouts uses "internal" directories that do not require permissions.

Note: CircledImageView.setForeground() listed as added in SDK 23
Compile possible with compile SDK 22 though, seem strange that this occurs. However, few Wear should use SDK22 anyway.