Repo and Travis tweaks
Tweaks to Travis setup. The intention was to try to limit the build time, but most time is due to the tools/platforms-tools dependencies. Some tweaks to caches, repos at least.

Specify play versions for Wear app
Use latest Play Services for phone app. This eliminates some Minify warnings. As just play-services-wearable is specified, the app size is not inflated, as well as "multi-dex" issues with later play services.
The previous version has some "FloatMath" suppressed warnings that could be fatal...

Use Maven repo for GraphView. This eliminates some Minify warnings with later releases too.
The GraphView version is slightly updated from 3.1.0 to 3.1.4 as 3.1.0 is not available in the Maven repo. Minor changes done.
The submodule can now be removed, not used. I see no reason to customize GraphView, so no real need for the subrepo.