#440 dropdown list are closed at updateView
An attempt fixing the issue, not completely working
It works OK if another issue is selected in the list
However, if same or none is selected, there is no update until another is selected.
If the GPS status as not connected before, the view will never be updated.
There must at least be an update if the status changes

Also lint removals