Restructure of Settings
The number of options have grown over time, I am to add some more (related to new senors and recording).
This is an attempt to group them logically in separate PreferenceScreens. Could have been categories but then the list will be long still.

MockHR and Eperimental should be moved to the menu in ConfigureHR

Added a few strings and modified other. There are other changes that need to be synced to Transifex

Some lint elimination