Use the appropriate parameters for requesting location updates
This commit makes GpsStatus use the user-chosen values for minTime and
minDistance required for GPS location updates, just as the TrackerGPS
code does. This saves a ton of battery life because we're not asking the
phone to keep its radios powered on constantly. I tested the code with
and without this patch using a virtual device and debug statements and
confirmed that this reduces the frequency of location updates.