Lint: FrameLayout can be replaced with <merge> tag
If a <FrameLayout> is the root of a layout and does not provide background or padding etc, it can often be replaced with a <merge> tag which is slightly more efficient. Note that this depends on context, so make sure you understand how the <merge> tag works before proceeding.  http://android-developers.blogspot.com/2009/03/android-layout-tricks-3-optimize-by.html