Lint: Set spinner distance unit with value
Consistence use of distance unit close to value. In Distance TitleSpinners, the value was in the title, hardcoded to (m), most everwhere else with the value. This change allow presenting localized distances eventually (and km instead of m where appropriate).
The second line titles were removed to improve grouping in the overview. For instance in Interval, it was previously two lines like the following
  Intervall type                                Time
  Intervall time (HH:MM:SS)          00:01:00
Is now
  Intervall type                                 Time
                                                         00:01:00
So less repeat text, remove HH:MM:SS that should be translated

TODO Use the numeric value from Spinner, avoid parsing again