Lint: SuppressLint("ClickableViewAccessibility")
Code seem OK to me, not seeing that performClick() is beneficial.
Suppressingseem to be the most common action.

--
../../src/org/runnerup/widget/NumberPicker.java:231: org/runnerup/widget/NumberPicker$5#onTouch should call View#performClick when a click is detected
 228     private final OnTouchListener buttonLongTouchListener = new OnTouchListener() {
 229         @Override
 230         //@SuppressLint("ClickableViewAccessibility")

 231         public boolean onTouch(View v, MotionEvent event) {

 232             if (event.getAction() == MotionEvent.ACTION_UP &&
 233                     ((longInc && v == incButton) ||
 234                     (longDec && v == decButton))) {
If a View that overrides onTouchEvent or uses an OnTouchListener does not also implement performClick and call it when clicks are detected, the View may not handle accessibility actions properly. Logic handling the click actions should ideally be placed in View#performClick as some accessibility services invoke performClick when a click action should occur.

To suppress this error, use the issue id "ClickableViewAccessibility" as explained in the Suppressing Warnings and Errors section.