Remove @TargetApi() annotations
This is used by Lint to set the minimal version when running Lint regardless of the version in Gradle. This is not needed as Lint runs separately for Latest and Froyo builds.
The annotation is mostly confusing and too broad (not overriding where needed).
In cases where the use level is raised (like Wear >= 4.3, some BLE code, MapBox etc) a version higher than minimum version can hide NewApi issues.
In some cases the level was set too low (like Froyo in latest), potentially giving warnings that are irrelevant.

The annotation was required here:
  \hrdevice\src\org\runnerup\hr\AndroidBLEHRProvider.java
  \app\src\org\runnerup\workout\feedback\RUTextToSpeech.java