Fix manually creating of workouts
This patch fixes a problem that had the visible effect that
it was not possible to manually create workouts, the workouts
where empty.

The root cause of this was a change in TitleSpinner that
removed the feature of having actual text as value.

It's unclear if that bug was visible somewhere else.