Remove Runnerup alpha
In 2.0 development builds have been published as runnerup.alpha, so it could be
installed together with the Play version.
This had a few side effects, especially for the Froyo build,
where "Up" navigation in the navigation menu failed (Back is OK).

As a replacement, debug builds set a ".debug" suffix so it can be installed with
release builds.