Step complettion: Use current point if closer to the target
Previously, the target were reached only when exceeding the target.
So if distance was 1000 m, the step was only completed if over 1000m
Similar for time, so average could be 1:01.

Time/distance targets are now handled similar to autolap changed recently,
so the current point is used if it is expected to be closer to the target
than next point, based on the previous point.