Permissions for activity and background for Android 10
If permissions are denied, give motivation and let the user try again
(unless "don't ask again" is ticked)

Remove snackbar as it will not rerequest permissions if the user ticks
"don't ask again".
Instead use a popup that asks the user to go to system settings,
without starting the workout. (Linking to system settings is not
recommended in the guidelines.)

TODO onRequestPermissionsResult() is not called, why permissions are
checked onCreate (similar to how requested for autostart GPS)
instead of when selecting Enable GPS.