make sure the time is correct even across DST changes
This change is a little larger than the last, but it should now be
easier to understand and it should keep the time as set even if the
entered date is not "today" and actually uses a different time zone.
This can happen, e.g., when the entered date is within DST and "today"
is not, or vise versa.