Update Formatter.java to fix month display on some languages (#1090)
There's an issue in the ICU notation format that makes some languages to look bad in the history screen of runner up, for example in catalan the history appears as "de gen. 2022" (there's an extra article) instead of "gen. 2022" that would be the proper text in this case.
Using LLL instead of MMM fixes this for the affected languages (Belarusian, Catalan, Croatian, Finnish, Greek, Lithuanian, Polish, Russian,
Slovak, Ukrainian and several more...) without touching the rest. (see the explanation in here https://developer.android.com/reference/android/icu/text/SimpleDateFormat)