More fixes for Wear OS (based on play store feedback) (#1231)
1) Remove the "action bar" that was shown when
the app started...and that was incorrectly centered
(how to fix that??). This was a complaint from play console.
This is fixed in manifest + styles.

2) Add the OngoingActivity concept to the Notification.
This does nothing on my old watch, but seems to work on
never versions (tried in emulator)

* Improve notification handling of wear app

Use non-deprecated api for read/write of data items.