Remove <application> tag from hrdevice manifest
The hrdevice module is a library and should not define its own
<application> tag. This caused manifest merger conflicts, specifically
using the hrdevice module's "@string/app_name" instead of the
common module's version. This resulted in the app being labeled
incorrectly as "HRDevice" instead of "RunnerUp".

Removing the tag from the library module resolves the conflict and ensures
the correct application name is used.