Deprecate sportEntries (string-array) in favor of static map in Sport (#1257)
This patchs adds a map between DB.ACTIVITY.SPORT_xxx and
R.strings.SPORT_xxx that is intended to be used instead of
array.sport_entries, which apparently is hard to handle
wrt to translations.

The patch currently use the existing sport_entries as bases
so that languages that has not yet added R.strings.SPORT_xxx
will be unaffected.