Graph-x-axis
This patch makes it possible to use time (instead of distance)
as X-axis in graphs...and makes it modifyable.
It also removes pace graph if there is no pace data.

Both of these features are planned for sport without gps,
i.e Treadmill. But in theory, one can imagine useing time instead
of distance as x-axis also for e.g. running. Maybe it should be
a switch ? (in a future version!)